requirements.txt
Flask>=2.0
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
alembic>=1.10
python-dateutil>=2.8
Flask-WTF>=1.0.0
pytest>=7.0
